% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-data.R
\name{fetch_census_demo}
\alias{fetch_census_demo}
\title{Census API call: demographics by county}
\usage{
fetch_census_demo()
}
\description{
info on race variable here https://api.census.gov/data/2019/pep/charagegroups/variables/RACE.json
}
