% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-data.R
\name{fetch_census_income}
\alias{fetch_census_income}
\title{Census API call: median income & pct in poverty by county}
\usage{
fetch_census_income()
}
\description{
Census API call: median income & pct in poverty by county
}
