% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-data.R
\name{tibblefy_census_html}
\alias{tibblefy_census_html}
\title{Behind the scenes function to convert raw data from census to a tibble}
\usage{
tibblefy_census_html(html)
}
\description{
Works for each of the tidy_census functions
}
