% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-data.R
\name{tidy_census_demo}
\alias{tidy_census_demo}
\title{Converts raw data from Census to a tidy df}
\usage{
tidy_census_demo(census_demo)
}
\description{
Cleans up some variables
}
\details{
To do: make df wide?
}
