% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-data.R
\name{tidy_census_income}
\alias{tidy_census_income}
\title{Converts raw data from census into a tidy df.
Adjusts colnames, converts char to numeric, tidies county info}
\usage{
tidy_census_income(census_income)
}
\description{
Converts raw data from census into a tidy df.
Adjusts colnames, converts char to numeric, tidies county info
}
